requirements.txt
attrs
selenium
